requirements.txt
pyvespa
datasets
pandas
docker
